+++ tests/end_to_end/test_suites/task_runner_tests.py
@@ -7,7 +7,6 @@ import logging
 from tests.end_to_end.utils.tr_common_fixtures import (
     fx_federation_tr,
     fx_federation_tr_dws,
-    fx_federation_tr_gandlf,
 )
 from tests.end_to_end.utils import federation_helper as fed_helper
 
@@ -55,28 +54,4 @@ def test_federation_via_dockerized_workspace(request, fx_federation_tr_dws):
     ), "Federation completion failed"
 
     best_agg_score = fed_helper.get_best_agg_score(fx_federation_tr_dws.aggregator.tensor_db_file)
-    log.info(f"Model best aggregated score post {request.config.num_rounds} is {best_agg_score}")
-
-
-@pytest.mark.task_runner_basic_gandlf
-def test_federation_via_native_gandlf(request, fx_federation_tr_gandlf):
-    """
-    Test federation via native task runner with GaNDLF.
-    IMPORTANT - ensure that all the pre-requisites steps for GanDLF are met before this test is run.
-    Refer file .github/workflows/gandlf.yaml for the same.
-    Args:
-        request (Fixture): Pytest fixture
-        fx_federation_tr (Fixture): Pytest fixture for native task runner
-    """
-    # Start the federation
-    assert fed_helper.run_federation(fx_federation_tr_gandlf)
-
-    # Verify the completion of the federation run
-    assert fed_helper.verify_federation_run_completion(
-        fx_federation_tr_gandlf,
-        test_env=request.config.test_env,
-        num_rounds=request.config.num_rounds,
-    ), "Federation completion failed"
-
-    best_agg_score = fed_helper.get_best_agg_score(fx_federation_tr_gandlf.aggregator.tensor_db_file)
     log.info(f"Model best aggregated score post {request.config.num_rounds} is {best_agg_score}")
