+++ tests/end_to_end/utils/tr_common_fixtures.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from tests.end_to_end.utils.tr_workspace import create_tr_workspace, create_tr_dws_workspace, create_tr_workspace_gandlf
+from tests.end_to_end.utils.tr_workspace import create_tr_workspace, create_tr_dws_workspace
 
 
 @pytest.fixture(scope="function")
@@ -21,23 +21,6 @@ def fx_federation_tr(request):
     """
     request.config.test_env = "task_runner_basic"
     return create_tr_workspace(request)
-
-
-@pytest.fixture(scope="function")
-def fx_federation_tr_gandlf(request):
-    """
-    Fixture for federation in case of GANDLF model. This fixture is used to create the model owner, aggregator, and collaborators.
-    It also creates workspace.
-    Assumption: OpenFL workspace is present for the model being tested.
-    Args:
-        request: pytest request object. Model name is passed as a parameter to the fixture from test cases.
-    Returns:
-        federation_details: Named tuple containing the objects for model owner, aggregator, and collaborators
-
-    Note: As this is a function level fixture, thus no import is required at test level.
-    """
-    request.config.test_env = "task_runner_basic_gandlf"
-    return create_tr_workspace_gandlf(request)
 
 
 @pytest.fixture(scope="function")
